+++ src/types/cli.d.ts
@@ -18,10 +18,12 @@ type CliParameters = {
   password: string;
   db: string;
   dbEngine?: string;
+  tableEngine?: string;
   timeout?: string;
   caCert?: string;
   cert?: string;
   key?: string;
+  abortDivergent?: boolean | string;
 };
 
 type QueryError = {